youtube_transcript_pipeline/requirements.txt
python-dotenv
youtube-transcript-api
yt-dlp
google-api-python-client
openai-whisper

PyYAML>=6.0.0,<7.0.0
openai>=1.0.0,<2.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg[binary]>=3.1.0,<4.0.0
pytest>=8.0.0,<9.0.0
